get rid of some warnings during build (#214)
(comparison between signed and unsigned integer expressions)

+++ src/knx/memory.cpp
@@ -127,7 +127,7 @@ void Memory::readMemory()
 void Memory::writeMemory()
 {
     // first get the necessary size of the writeBuffer
-    size_t writeBufferSize = _metadataSize;
+    uint16_t writeBufferSize = _metadataSize;
     for (int i = 0; i < _saveCount; i++)
         writeBufferSize = MAX(writeBufferSize, _saveRestores[i]->saveSize());
